Abstract the logic that makes HTTP requests during testing to an isolated class to prevent confusion when chaining headers/cookies
